content/057/services/catalog/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic==2.8.2
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-pymongo==0.47b0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2
